ui: rename app_stats to appStats
Release note: None
