+++ pkg/ui/src/util/appStats.ts
